pyproject.toml
[project]
name = "classroom-transcripts"
version = "0.1.0"
description = "Classroom Transcripts for Coaching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "assemblyai>=0.37.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
    "azure-storage-blob>=12.24.1",
    "azure-identity>=1.19.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.160.0",
    "plotly>=6.0.0",
    "scipy>=1.15.1",
    "numpy>=2.2.2",
    "azure-data-tables>=12.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/classroom_transcripts"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --doctest-modules"
testpaths = [
    "tests",
    "src/functions/tests",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "ruff>=0.2.1",
    "watchdog>=6.0.0",
]

[tool.uv.workspace]
members = ["src/functions"]
